Module twitch_api2::eventsub::channel

This is supported on crate feature eventsub only.

Subscription types regarding channels

Modules

ban

A viewer is banned from the specified channel.

channel_points_custom_reward

Custom channel points rewards on specific channel has been changed, removed or updated.

channel_points_custom_reward_redemption

A viewer has redeemed a custom channel points reward or a redemption of a channel points custom reward has been updated for the specified channel.

cheer

A user cheers on the specified channel.

follow

A specified channel receives a follow.

hypetrain

A hype train has started, progressed or ended.

poll

Poll on a specific channel has been begun, ended or progressed.

prediction

Prediction on the specified channel begins, progresses, locks or ends.

raid

A a broadcaster raids another broadcaster’s channel.

subscribe

A specified channel receives a subscriber. This does not include resubscribes.

subscription

Subscription on a specified channel has changed

unban

A viewer is unbanned from the specified channel.

update

Channel has updated the category, title, mature flag, or broadcast language.

Structs

ChannelBanV1

channel.ban: a viewer is banned from the specified channel.

ChannelBanV1Payload

channel.ban response payload.

ChannelCheerV1

channel.cheer: a user cheers on the specified channel.

ChannelCheerV1Payload

channel.cheer response payload.

ChannelFollowV1

channel.follow: a specified channel receives a follow.

ChannelFollowV1Payload

channel.follow response payload.

ChannelHypeTrainBeginV1

channel.hype_train.begin: a hype train begins on the specified channel.

ChannelHypeTrainBeginV1Payload

channel.hype_train.begin response payload.

ChannelHypeTrainEndV1

channel.hype_train.end: a hype train ends on the specified channel.

ChannelHypeTrainEndV1Payload

channel.hype_train.end response payload.

ChannelHypeTrainProgressV1

channel.hype_train.progress: a hype train makes progress on the specified channel.

ChannelHypeTrainProgressV1Payload

channel.hype_train.progress response payload.

ChannelPointsCustomRewardAddV1

channel.channel_points_custom_reward.add: a custom channel points reward has been created for the specified channel.

ChannelPointsCustomRewardAddV1Payload

channel.channel_points_custom_reward.add response payload.

ChannelPointsCustomRewardRedemptionAddV1

channel.channel_points_custom_reward_redemption.add: a viewer has redeemed a custom channel points reward on the specified channel.

ChannelPointsCustomRewardRedemptionAddV1Payload

channel.channel_points_custom_reward_redemption.add response payload.

ChannelPointsCustomRewardRedemptionUpdateV1

channel.channel_points_custom_reward_redemption.update: a redemption of a channel points custom reward has been updated for the specified channel.

ChannelPointsCustomRewardRedemptionUpdateV1Payload

channel.channel_points_custom_reward_redemption.update response payload.

ChannelPointsCustomRewardRemoveV1

channel.channel_points_custom_reward.remove: a custom channel points reward has been removed from the specified channel.

ChannelPointsCustomRewardRemoveV1Payload

channel.channel_points_custom_reward.remove response payload.

ChannelPointsCustomRewardUpdateV1

channel.channel_points_custom_reward.update: a custom channel points reward has been updated for the specified channel.

ChannelPointsCustomRewardUpdateV1Payload

channel.channel_points_custom_reward.update response payload.

ChannelPollBeginV1

channel.poll.begin: a poll begins on the specified channel.

ChannelPollBeginV1Payload

channel.poll.begin response payload.

ChannelPollEndV1

channel.hype_train.end: a poll ends on the specified channel.

ChannelPollEndV1Payload

channel.hype_train.end response payload.

ChannelPollProgressV1

channel.hype_train.progress: an user responds to a poll on the specified channel

ChannelPollProgressV1Payload

channel.hype_train.progress response payload.

ChannelPredictionBeginV1

channel.prediction.begin: a Prediction begins on the specified channel

ChannelPredictionBeginV1Payload

channel.prediction.begin response payload.

ChannelPredictionEndV1

channel.hype_train.end: a Prediction ends on the specified channel.

ChannelPredictionEndV1Payload

channel.hype_train.end response payload.

ChannelPredictionLockV1

channel.hype_train.lock: an user responds to a prediction on the specified channel

ChannelPredictionLockV1Payload

channel.hype_train.lock response payload.

ChannelPredictionProgressV1

channel.hype_train.progress: an user responds to a prediction on the specified channel

ChannelPredictionProgressV1Payload

channel.hype_train.progress response payload.

ChannelRaidV1

channel.raid: a a broadcaster raids another broadcaster’s channel.

ChannelRaidV1Payload

channel.raid response payload.

ChannelSubscribeV1

channel.subscribe: a specified channel receives a subscriber. This does not include resubscribes.

ChannelSubscribeV1Payload

channel.subscribe response payload.

ChannelSubscriptionEndV1

channel.subscription.end: a subscription to the specified channel expires.

ChannelSubscriptionEndV1Payload

channel.subscription.end response payload.

ChannelSubscriptionGiftV1

channel.subscription.gift: a subscription to the specified channel expires.

ChannelSubscriptionGiftV1Payload

channel.subscription.gift response payload.

ChannelSubscriptionMessageV1

channel.subscription.message: a subscription to the specified channel expires.

ChannelSubscriptionMessageV1Payload

channel.subscription.message response payload.

ChannelUnbanV1

channel.unban: a viewer is unbanned from the specified channel.

ChannelUnbanV1Payload

channel.unban response payload.

ChannelUpdateV1

channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelUpdateV1Payload

channel.update response payload.