Struct twitch_api2::eventsub::channel::channel_points_custom_reward::remove::ChannelPointsCustomRewardRemoveV1Payload

#[non_exhaustive]pub struct ChannelPointsCustomRewardRemoveV1Payload {
    pub background_color: String,
    pub broadcaster_user_id: UserId,
    pub broadcaster_user_name: UserName,
    pub cooldown_expires_at: Option<Timestamp>,
    pub cost: i64,
    pub default_image: Option<Image>,
    pub global_cooldown: GlobalCooldown,
    pub id: RewardId,
    pub image: Option<Image>,
    pub is_enabled: bool,
    pub is_in_stock: bool,
    pub is_paused: bool,
    pub is_user_input_required: bool,
    pub max_per_stream: Max,
    pub max_per_user_per_stream: Max,
    pub prompt: String,
    pub redemptions_redeemed_current_stream: Option<u32>,
    pub should_redemptions_skip_request_queue: bool,
    pub title: String,
}

This is supported on crate feature eventsub only.

channel.channel_points_custom_reward.remove response payload.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct {{ .. }} syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
background_color: String

Custom background color for the reward. Format: Hex with # prefix. Example: #FA1ED2.

broadcaster_user_id: UserId

The requested broadcaster ID.

broadcaster_user_name: UserName

The requested broadcaster name.

cooldown_expires_at: Option<Timestamp>

Timestamp of the cooldown expiration. null if the reward isn’t on cooldown.

cost: i64

The reward cost.

default_image: Option<Image>

Set of default images of 1x, 2x and 4x sizes for the reward.

global_cooldown: GlobalCooldown

Whether a cooldown is enabled and what the cooldown is in seconds.

id: RewardId

The reward identifier.

image: Option<Image>

Set of custom images of 1x, 2x and 4x sizes for the reward. Can be null if no images have been uploaded.

is_enabled: bool

Is the reward currently enabled. If false, the reward won’t show up to viewers.

is_in_stock: bool

Is the reward currently in stock. If false, viewers can’t redeem.

is_paused: bool

Is the reward currently paused. If true, viewers can’t redeem.

is_user_input_required: bool

Does the viewer need to enter information when redeeming the reward.

max_per_stream: Max

Whether a maximum per stream is enabled and what the maximum is.

max_per_user_per_stream: Max

Whether a maximum per user per stream is enabled and what the maximum is.

prompt: String

The reward description.

redemptions_redeemed_current_stream: Option<u32>

The number of redemptions redeemed during the current live stream. Counts against the max_per_stream limit. null if the broadcasters stream isn’t live or max_per_stream isn’t enabled.

should_redemptions_skip_request_queue: bool

Should redemptions be set to fulfilled status immediately when redeemed and skip the request queue instead of the normal unfulfilled status.

title: String

The reward title.