Module twitch_api2::eventsub::channel::channel_points_custom_reward_redemption

This is supported on crate feature eventsub only.

A viewer has redeemed a custom channel points reward or a redemption of a channel points custom reward has been updated for the specified channel.

Modules

add

A viewer has redeemed a custom channel points reward on the specified channel.

update

A redemption of a channel points custom reward has been updated for the specified channel.

Structs

ChannelPointsCustomRewardRedemptionAddV1

channel.channel_points_custom_reward_redemption.add: a viewer has redeemed a custom channel points reward on the specified channel.

ChannelPointsCustomRewardRedemptionAddV1Payload

channel.channel_points_custom_reward_redemption.add response payload.

ChannelPointsCustomRewardRedemptionUpdateV1

channel.channel_points_custom_reward_redemption.update: a redemption of a channel points custom reward has been updated for the specified channel.

ChannelPointsCustomRewardRedemptionUpdateV1Payload

channel.channel_points_custom_reward_redemption.update response payload.

Reward

Basic information about the reward that was redeemed, at the time it was redeemed.

Enums

RedemptionStatus

Custom reward redemption statuses: UNFULFILLED, FULFILLED or CANCELED