Module twitch_api2::eventsub::channel::hypetrain

This is supported on crate feature eventsub only.

A hype train has started, progressed or ended.

Modules

begin
A hype train begins on the specified channel.

end
A hype train ends on the specified channel.

progress
A hype train makes progress on the specified channel.

Structs

ChannelHypeTrainBeginV1
channel.hype_train.begin: a hype train begins on the specified channel.

ChannelHypeTrainBeginV1Payload
channel.hype_train.begin response payload.

ChannelHypeTrainEndV1
channel.hype_train.end: a hype train ends on the specified channel.

ChannelHypeTrainEndV1Payload
channel.hype_train.end response payload.

ChannelHypeTrainProgressV1
channel.hype_train.progress: a hype train makes progress on the specified channel.

ChannelHypeTrainProgressV1Payload
channel.hype_train.progress response payload.

Contribution
A contribution to hype train

Enums

ContributionType
Type of contribution