Module twitch_api2::eventsub::channel::prediction

This is supported on crate feature eventsub only.

Prediction on the specified channel begins, progresses, locks or ends.

Modules

begin

A Prediction begins on the specified channel

end

A Prediction ends on the specified channel.

lock

A user responds to a prediction on the specified channel

progress

A user responds to a prediction on the specified channel

Structs

ChannelPredictionBeginV1

channel.prediction.begin: a Prediction begins on the specified channel

ChannelPredictionBeginV1Payload

channel.prediction.begin response payload.

ChannelPredictionEndV1

channel.hype_train.end: a Prediction ends on the specified channel.

ChannelPredictionEndV1Payload

channel.hype_train.end response payload.

ChannelPredictionLockV1

channel.hype_train.lock: an user responds to a prediction on the specified channel

ChannelPredictionLockV1Payload

channel.hype_train.lock response payload.

ChannelPredictionProgressV1

channel.hype_train.progress: an user responds to a prediction on the specified channel

ChannelPredictionProgressV1Payload

channel.hype_train.progress response payload.