Module twitch_api2::eventsub::channel::update

This is supported on crate feature eventsub only.

Channel has updated the category, title, mature flag, or broadcast language.

Structs

ChannelUpdateV1
channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelUpdateV1Payload
channel.update response payload.