Module twitch_api2::eventsub::user

This is supported on crate feature eventsub only.

Subscription types regarding users

Modules

authorization
A user has revoked authorization for your client id.

update
Specified user updates their account.

Structs

UserAuthorizationRevokeV1
user.authorization.revoke: a user has revoked authorization for your client id. Use this webhook to meet government requirements for handling user data, such as GDPR, LGPD, or CCPA.

UserAuthorizationRevokeV1Payload
user.authorization.revoke response payload.

UserUpdateV1
user.update: user updates their account.

UserUpdateV1Payload
user.update response payload.