Module twitch_api2::eventsub::user::authorization

This is supported on crate feature eventsub only.

A user has revoked authorization for your client id.

Structs

UserAuthorizationRevokeV1
user.authorization.revoke: a user has revoked authorization for your client id. Use this webhook to meet government requirements for handling user data, such as GDPR, LGPD, or CCPA.

UserAuthorizationRevokeV1Payload
user.authorization.revoke response payload.