Module twitch_api2::helix

Helix endpoints or the New Twitch API

Re-exports

pub use twitch_oauth2::Scope;

Modules

channels
Endpoints regarding channels

clips
Endpoints regarding clips

moderation
Endpoints regarding moderation

streams
Endpoints regarding streams

subscriptions
Endpoints regarding subscriptions

users
Endpoints regarding users

Structs

HelixClient	feature="helix"
Client for Helix or the New Twitch API

Pagination
A cursor for pagination. This is needed because of how pagination is represented in the New Twitch API

Response
Response retrieved from endpoint. Data is the type in Request::Response

Enums

Error
Errors from the query serializer

RequestError
Errors for HelixClient::req_get and similar functions.

Traits

Paginated
Request can be paginated with a cursor

Request
A request is a Twitch endpoint, see New Twitch API reference

RequestGet
Helix endpoint GETs information

RequestPatch
Helix endpoint PATCHs information

RequestPost
Helix endpoint POSTs information

Functions

repeat_query
Repeat url query items with name

Type Definitions

Cursor
A cursor is a pointer to the current "page" in the twitch api pagination