Module twitch_api2::helix::bits

This is supported on crate feature helix only.

Helix endpoints regarding bits

Examples

let client = HelixClient::default();
let req = GetCheermotesRequest::builder()
    .broadcaster_id(Some("1234".into()))
    .build();

println!("{:?}", &client.req_get(req, &token).await?.data);

Modules

get_bits_leaderboard

Gets a ranked list of Bits leaderboard information for an authorized broadcaster. get-bits-leaderboard

get_cheermotes

Retrieves the list of available Cheermotes, animated emotes to which viewers can assign Bits, to cheer in chat. get-cheermotes

Structs

BitsLeaderboard

Return Values for Get Bits Leaderboard

Cheermote

Return Values for Get Cheermotes

GetBitsLeaderboardRequest

Query Parameters for Get Bits Leaderboard

GetCheermotesRequest

Query Parameters for Get Cheermotes