Module twitch_api2::helix::bits::get_cheermotes

This is supported on crate feature helix only.

Retrieves the list of available Cheermotes, animated emotes to which viewers can assign Bits, to cheer in chat. get-cheermotes

Structs

Cheermote
Return Values for Get Cheermotes

CheermoteImage
URLs to the same image, animated or static.

CheermoteImageArray
Images in sizes

CheermoteImages
Images for different themes

GetCheermotesRequest
Query Parameters for Get Cheermotes

Level
Level of cheermote image

Tiers
Information about tier of emote

Enums

CheermoteType
Types of cheermotes