Module twitch_api2::helix::chat::get_channel_emotes

This is supported on crate feature helix only.

Gets all custom emotes for a specific Twitch channel including subscriber emotes, Bits tier emotes, and follower emotes. get-channel-emotes

Accessing the endpoint

Request: GetChannelEmotesRequest

To use this endpoint, construct a GetChannelEmotesRequest with the GetChannelEmotesRequest::builder() method.

use twitch_api2::helix::chat::get_channel_emotes;
let request = get_channel_emotes::GetChannelEmotesRequest::builder()
    .broadcaster_id("1234".to_string())
    .build();

Response: ChannelEmote

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, chat::get_channel_emotes};
let request = get_channel_emotes::GetChannelEmotesRequest::builder()
    .broadcaster_id("1234".to_string())
    .build();
let response: Vec<helix::chat::ChannelEmote> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetChannelEmotesRequest::parse_response(None, &request.get_uri(), response)

Structs

GetChannelEmotesRequest

Query Parameters for Get Channel Emotes

Type Definitions

GetChannelEmotesResponse

Return Values for Get Channel Emotes