Module twitch_api2::helix::chat::get_emote_sets

This is supported on crate feature helix only.

Gets all Twitch emotes for one or more specific emote sets. get-emote-sets

Accessing the endpoint

Request: GetEmoteSetsRequest

To use this endpoint, construct a GetEmoteSetsRequest with the GetEmoteSetsRequest::builder() method.

use twitch_api2::helix::chat::get_emote_sets;
let request = get_emote_sets::GetEmoteSetsRequest::builder()
    .emote_set_id(vec!["1234".into()])
    .build();

Response: Emote

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, chat::get_emote_sets};
let request = get_emote_sets::GetEmoteSetsRequest::builder()
    .emote_set_id(vec!["1234".into()])
    .build();
let response: Vec<helix::chat::get_emote_sets::Emote> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetEmoteSetsRequest::parse_response(None, &request.get_uri(), response)

Structs

Emote

Return Values for Get Channel Emotes

GetEmoteSetsRequest

Query Parameters for Get Channel Emotes