Struct twitch_api2::helix::eventsub::get_eventsub_subscriptions::GetEventSubSubscriptionsRequest

#[non_exhaustive]
pub struct GetEventSubSubscriptionsRequest {
    pub status: Option<Status>,
    pub after: Option<Cursor>,
    pub first: Option<usize>,
}

This is supported on crate features helix and eventsub only.

Query Parameters for Get EventSub Subscriptions

get-eventsub-subscriptions

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct {{ .. }} syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
status: Option<Status>

Include this parameter to filter subscriptions by their status.

after: Option<Cursor>

Cursor for forward pagination

first: Option<usize>

Maximum number of objects to return. Maximum: 100. Default: 20.

Implementations

impl GetEventSubSubscriptionsRequest

pub fn builder() -> GetEventSubSubscriptionsRequestBuilder<((), (), ())>

Create a builder for building GetEventSubSubscriptionsRequest. On the builder, call .status(...)(optional), .after(...)(optional), .first(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of GetEventSubSubscriptionsRequest.