Module twitch_api2::helix::moderation

This is supported on crate feature helix only.

Helix endpoints regarding moderation

Modules

check_automod_status

Determines whether a string message meets the channel’s AutoMod requirements. check-automod-status

get_banned_events

Returns all banned and timed-out users in a channel. get-banned-events

get_banned_users

Returns all banned and timed-out users in a channel. get-banned-users

get_moderator_events

Returns a list of moderators or users added and removed as moderators from a channel. get-moderator-events

get_moderators

Returns all moderators in a channel. get-moderators

manage_held_automod_messages

Allow or deny a message that was held for review by AutoMod. manage-held-automod-messages

Structs

BannedEvent

Return Values for Get Banned Events

BannedUser

Return Values for Get Banned Users

CheckAutoModStatus

Return Values for Check AutoMod Status

CheckAutoModStatusBody

Body Parameters for Check AutoMod Status

CheckAutoModStatusRequest

Query Parameters for Check AutoMod Status

GetBannedEventsRequest

Query Parameters for Get Banned Events

GetBannedUsersRequest

Query Parameters for Get Banned Users

GetModeratorEventsRequest

Query Parameters for Get Moderators Events

GetModeratorsRequest

Query Parameters for Get Moderators

ManageHeldAutoModMessagesBody

Body Parameters for Manage Held AutoMod Messages

ManageHeldAutoModMessagesRequest

Query Parameters for Manage Held AutoMod Messages

Moderator

Return Values for Get Moderators

ModeratorEvent

Return Values for Get Moderators Events

Enums

AutoModAction

Action to take for a message.

ManageHeldAutoModMessages

Return Values for Manage Held AutoMod Messages