Module twitch_api2::helix::moderation::get_moderator_events

This is supported on crate feature helix only.

Returns a list of moderators or users added and removed as moderators from a channel. get-moderator-events

Accessing the endpoint

Request: GetModeratorEventsRequest

To use this endpoint, construct a GetModeratorEventsRequest with the GetModeratorEventsRequest::builder() method.

use twitch_api2::helix::moderation::get_moderator_events;
let request = get_moderator_events::GetModeratorEventsRequest::builder()
    .broadcaster_id("1234")
    .build();

Response: ModeratorEvent

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, moderation::get_moderator_events};
let request = get_moderator_events::GetModeratorEventsRequest::builder()
    .broadcaster_id("1234")
    .build();
let response: Vec<get_moderator_events::ModeratorEvent> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with request.parse_response(&request.get_uri()?)

Structs

GetModeratorEventsRequest
Query Parameters for Get Moderators Events

ModeratorEvent
Return Values for Get Moderators Events