Module twitch_api2::helix::points

This is supported on crate feature helix only.

Helix endpoints regarding channel points/redeems

Examples

let client = HelixClient::new();
let request = GetCustomRewardRedemptionRequest::builder()
    .broadcaster_id("274637212".to_string())
    .reward_id("92af127c-7326-4483-a52b-b0da0be61c01".to_string())
    .status(CustomRewardRedemptionStatus::Canceled)
    .build();

println!("{:?}", &client.req_get(request, &token).await?.data.get(0));

Modules

get_custom_reward_redemption
Returns Custom Reward Redemption objects for a Custom Reward on a channel that was created by the same client_id.

update_redemption_status
Updates the status of Custom Reward Redemption objects on a channel that are in the UNFULFILLED status.

Structs

CustomRewardRedemption
Return Values for Get Custom Reward Redemption

GetCustomRewardRedemptionRequest
Query Parameters for Get Custom Reward Redemption

UpdateRedemptionStatusBody
Body Parameters for Update Redemption Status

UpdateRedemptionStatusRequest
Query Parameters for Update Redemption Status

Enums

CustomRewardRedemptionStatus
Custom reward redemption statuses: UNFULFILLED, FULFILLED or CANCELED

UpdateRedemptionStatusInformation
FIXME: Returns an object. Return Values for Update Redemption Status