Struct twitch_api2::helix::points::update_redemption_status::UpdateRedemptionStatusRequest

#[non_exhaustive]
pub struct UpdateRedemptionStatusRequest {
    pub broadcaster_id: UserId,
    pub reward_id: RewardId,
    pub id: RedemptionId,
}

This is supported on crate feature helix only.

Query Parameters for Update Redemption Status

update-redemption-status

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

broadcaster_id: UserId

Provided broadcaster_id must match the user_id in the auth token.

reward_id: RewardId

ID of the Custom Reward the redemptions to be updated are for.

id: RedemptionId

ID of the Custom Reward Redemption to update, must match a Custom Reward Redemption on broadcaster_id’s channel

Implementations

impl UpdateRedemptionStatusRequest

pub fn builder() -> UpdateRedemptionStatusRequestBuilder<((), (), ())>

Create a builder for building UpdateRedemptionStatusRequest. On the builder, call .broadcaster_id(...), .reward_id(...), .id(...) to set the values of the fields. Finally, call .build() to create the instance of UpdateRedemptionStatusRequest.