Module twitch_api2::helix::predictions::create_prediction

This is supported on crate feature helix only.

Get information about all Channel Points Predictions or specific Channel Points Predictions for a Twitch channel.

Results are ordered by most recent, so it can be assumed that the currently active or locked Prediction will be the first item. create-prediction

Accessing the endpoint

Request: CreatePredictionRequest

To use this endpoint, construct a CreatePredictionRequest with the CreatePredictionRequest::new() method.

use twitch_api2::helix::predictions::create_prediction;
let request = create_prediction::CreatePredictionRequest::new();

Body: CreatePredictionBody

We also need to provide a body to the request containing what we want to change.

let body = create_prediction::CreatePredictionBody::builder()
    .broadcaster_id("141981764")
    .title("Any leeks in the stream?")
    .outcomes(create_prediction::NewPredictionOutcome::new_tuple(
        "Yes, give it time.",
        "Definitely not.",
    ))
    .prediction_window(120)
    .build();

Response: CreatePredictionResponse

Send the request to receive the response with HelixClient::req_post().

use twitch_api2::helix::{self, predictions::create_prediction};
let request = create_prediction::CreatePredictionRequest::builder()
    .build();
let body = create_prediction::CreatePredictionBody::builder()
    .broadcaster_id("141981764")
    .title("Any leeks in the stream?")
    .outcomes(create_prediction::NewPredictionOutcome::new_tuple("Yes, give it time.", "Definitely not."))
    .prediction_window(120)
    .build();
let response: create_prediction::CreatePredictionResponse = client.req_post(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with CreatePredictionRequest::parse_response(None, &request.get_uri(), response)

Structs

CreatePredictionBody

Body Parameters for Create Prediction

CreatePredictionRequest

Query Parameters for Create Prediction

NewPredictionOutcome

Choice settings for a poll

Type Definitions

CreatePredictionResponse

Return Values for Create Prediction