Module twitch_api2::helix::predictions::get_predictions

This is supported on crate feature helix only.

Get information about all predictions or specific predictions for a Twitch channel. Prediction information is available for 90 days. get-predictions

Request: GetPredictionsRequest

To use this endpoint, construct a GetPredictionsRequest with the GetPredictionsRequest::builder() method.

use twitch_api2::helix::predictions::get_predictions;
let request = get_predictions::GetPredictionsRequest::builder()
    .id(vec!["ed961efd-8a3f-4cf5-a9d0-e616c590cd2a".into()])
    .broadcaster_id("1234")
    .build();

Response: Prediction

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, predictions::get_predictions};
let request = get_predictions::GetPredictionsRequest::builder()
    .id(vec!["ed961efd-8a3f-4cf5-a9d0-e616c590cd2a".into()])
    .broadcaster_id("1234")
    .build();
let response: Vec<get_predictions::Prediction> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetPredictionsRequest::parse_response(None, &request.get_uri(), response)

Re-exports

pub use types::PredictionOutcome;

pub use types::PredictionOutcomeId;

pub use types::PredictionStatus;

Structs

GetPredictionsRequest

Query Parameters for Get predictions

Prediction

Return Values for Get predictions