Module twitch_api2::helix::search::search_channels

This is supported on crate feature helix only.

Returns a list of channels (users who have streamed within the past 6 months) that match the query via channel name or description either entirely or partially. search-channels

Accessing the endpoint

Request: SearchChannelsRequest

To use this endpoint, construct a SearchChannelsRequest with the SearchChannelsRequest::builder() method.

use twitch_api2::helix::search::search_channels;
let request = search_channels::SearchChannelsRequest::builder()
    .query("hello")
    .build();

Response: Channel

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, search::search_channels};
let request = search_channels::SearchChannelsRequest::builder()
    .query("hello")
    .build();
let response: Vec<search_channels::Channel> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with SearchChannelsRequest::parse_response(None, &request.get_uri(), response)

Structs

Channel

Return Values for Search Channels

SearchChannelsRequest

Query Parameters for Search Channels