Module twitch_api2::helix::streams

This is supported on crate feature helix only.

Endpoints regarding streams

Examples

let client = HelixClient::new();
let req = GetStreamsRequest::builder()
    .user_login(vec!["justinfan1337".to_string()])
    .build();

// If this doesn't return a result, that would mean the stream is not live.
println!("{:?}", &client.req_get(req, &token).await?.data.get(0));

Re-exports

pub use get_stream_tags::GetStreamTagsRequest;

pub use get_stream_tags::Tag;

pub use get_streams::GetStreamsRequest;

pub use get_streams::Stream;

Modules

get_stream_tags
Gets the list of tags for a specified stream (channel). get-stream-tags

get_streams
Gets information about active streams. get-streams

Enums

StreamType
Gotten from Stream.type_