Trait twitch_api2::helix::Request

pub trait Request: Serialize {
    type Response;

    const PATH: &'static str;
    const SCOPE: &'static [Scope];
    const OPT_SCOPE: &'static [Scope];
    fn query(&self) -> Result<String, Error> { ... }
}

A request is a Twitch endpoint, see New Twitch API reference

Associated Types

type Response

Response type. twitch's response will deserialize to this.

Loading content...
Associated Constants

const PATH: &'static str

The path to the endpoint relative to the helix root. eg. channels for Get Channel Information

const SCOPE: &'static [Scope]

Scopes needed by this endpoint

const OPT_SCOPE: &'static [Scope]

Optional scopes needed by this endpoint

Loading content...
Provided methods

fn query(&self) -> Result<String, Error>

Defines layout of the url parameters.

Loading content...