Module twitch_api2::helix::users

This is supported on crate feature helix only.

Helix endpoints regarding users

Examples

let client = HelixClient::new();
let req = GetUsersRequest::builder()
    .login(vec!["justinfan1337".to_string()])
    .build();

println!("{:?}", &client.req_get(req, &token).await?.data);

Modules

block_user
Blocks the specified user on behalf of the authenticated user. block-user

create_user_follows
Adds a specified user to the followers of a specified channel. create-user-follows

delete_user_follows
Deletes a specified user from the followers of a specified channel. delete-user-follows

get_user_block_list
Gets a specified user’s block list. The list is sorted by when the block occurred in descending order (i.e. most recent block first). get-user-block-list

get_users
Gets information about one or more specified Twitch users. get-users

get_users_follows
Gets information on follow relationships between two Twitch users. get-users-follows

unblock_user
Unblocks the specified user on behalf of the authenticated user. unblock-user

Structs

BlockUserRequest
Query Parameters for Block User

CreateUserFollowsBody
Body Parameters for Create User Follows

CreateUserFollowsRequest
Query Parameters for Create User Follows

DeleteUserFollowsRequest
Query Parameters for Delete Users Follows

GetUserBlockListRequest
Query Parameters for Get Users Block List

GetUsersFollowsRequest
Query Parameters for Get Users Follows

GetUsersRequest
Query Parameters for Get Users

UnblockUserRequest
Query Parameters for Unblock User

User
Return Values for Get Users

UserBlock
Return Values for Get Users Block List

UsersFollow
Return Values for Get Users Follows

Enums

BlockUser
Return Values for Block User

CreateUserFollows
Return Values for Create User Follows

DeleteUserFollow
Return Values for Delete Users Follows

UnblockUser
Return Values for Unblock User