Module twitch_api2::helix::users::delete_user_follows

This is supported on crate feature helix only.

Deletes a specified user from the followers of a specified channel. delete-user-follows

Notes

This doesn’t seem to work for removing people who follow owner of token. Use Block User for that

Accessing the endpoint

Request: DeleteUserFollowsRequest

To use this endpoint, construct a DeleteUserFollowsRequest with the DeleteUserFollowsRequest::builder() method.

use twitch_api2::helix::users::delete_user_follows;
let request = delete_user_follows::DeleteUserFollowsRequest::builder()
    .from_id("1234").to_id("4321")
    .build();

Response: DeleteUserFollow

Send the request to receive the response with HelixClient::req_delete().

use twitch_api2::helix::{self, users::delete_user_follows};
let request = delete_user_follows::DeleteUserFollowsRequest::builder()
    .from_id("1234").to_id("4321")
    .build();
let response: delete_user_follows::DeleteUserFollow = client.req_delete(request, &token).await?;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with request.parse_response(&request.get_uri()?)

Structs

DeleteUserFollowsRequest
Query Parameters for Delete Users Follows

Enums

DeleteUserFollow
Return Values for Delete Users Follows