Module twitch_api2::helix::users::get_user_block_list

This is supported on crate feature helix only.

Gets a specified user’s block list. The list is sorted by when the block occurred in descending order (i.e. most recent block first). get-user-block-list

Request: GetUserBlockListRequest

To use this endpoint, construct a GetUserBlockListRequest with the GetUserBlockListRequest::builder() method.

use twitch_api2::helix::users::get_user_block_list;
let request = get_user_block_list::GetUserBlockListRequest::builder()
    .broadcaster_id("1234".to_string())
    .build();

Response: UserBlock

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, users::get_user_block_list};
let request = get_user_block_list::GetUserBlockListRequest::builder()
    .broadcaster_id("1234".to_string())
    .build();
let response: Vec<get_user_block_list::UserBlock> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetUserBlockListRequest::parse_response(None, &request.get_uri(), response)

Structs

GetUserBlockListRequest

Query Parameters for Get Users Block List

UserBlock

Return Values for Get Users Block List