Module twitch_api2::helix::users::get_users_follows

This is supported on crate feature helix only.

Gets information on follow relationships between two Twitch users. get-users-follows

Request: GetUsersFollowsRequest

To use this endpoint, construct a GetUsersFollowsRequest with the GetUsersFollowsRequest::builder() method.

use twitch_api2::helix::users::get_users_follows;
let request = get_users_follows::GetUsersFollowsRequest::builder()
    .to_id("1234".to_string())
    .build();

Response: UsersFollows

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, users::get_users_follows};
let request = get_users_follows::GetUsersFollowsRequest::builder()
    .to_id("1234".to_string())
    .build();
let response: Vec<get_users_follows::FollowRelationship> = client.req_get(request, &token).await?.data.follow_relationships;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetUsersFollowsRequest::parse_response(None, &request.get_uri(), response)

Structs

FollowRelationship
Describes a follow relationship

GetUsersFollowsRequest
Query Parameters for Get Users Follows

UsersFollows
Return Values for Get Users Follows