Struct twitch_api2::pubsub::hypetrain::Config

#[non_exhaustive]
pub struct Config {

    pub callout_emote_id: EmoteId,
    pub callout_emote_token: String,
    pub channel_id: Option<UserId>,
    pub conductor_rewards: ConductorRewards,
    pub cooldown_duration: i64,
    pub difficulty: HypeTrainDifficulty,
    pub difficulty_settings: HashMap<HypeTrainDifficulty, Vec<Level>>,
    pub is_enabled: bool,
    pub kickoff: Kickoff,
    pub level_duration: i64,
    pub notification_thresholds: NotificationThresholds,
    pub participation_conversion_rates: ParticipationConversionRates,
    pub theme_color: Option<String>,
    pub primary_hex_color: Option<String>,
    pub use_personalized_settings: Option<bool>,
    pub use_theme_color: Option<bool>,
    pub use_creator_color: Option<bool>,
    // some fields omitted

}

This is supported on crate features pubsub and unsupported only.

Configuration of hype train

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

callout_emote_id: EmoteId

Hype train public callout emote ID

callout_emote_token: String

Hype train public callout emote token

channel_id: Option<UserId>

ID of channel

conductor_rewards: ConductorRewards

Rewards for conductors

cooldown_duration: i64

Cooldown duration in nanoseconds for hype train

difficulty: HypeTrainDifficulty

Difficulty of hype train

difficulty_settings: HashMap<HypeTrainDifficulty, Vec<Level>>

Difficulty settings

is_enabled: bool

Whether or not Hype Train is enabled in channel

kickoff: Kickoff

Support events that must occur within a duration of time to kick off a Hype Train.

level_duration: i64

Duration in nanoseconds of each level

notification_thresholds: NotificationThresholds

Thresholds for notifications

participation_conversion_rates: ParticipationConversionRates

Conversion rates for participations

theme_color: Option<String>

Theme color of channel

None if use_theme_color is set to false

primary_hex_color: Option<String>

Primary color of hex

use_personalized_settings: Option<bool>

Uses personalized settings

use_theme_color: Option<bool>

Use theme color or not

use_creator_color: Option<bool>

Use creator color or not