Enum twitch_api2::pubsub::moderation::ModerationActionCommand

#[non_exhaustive]
pub enum ModerationActionCommand {

    Delete,
    Timeout,
    Untimeout,
    Mod,
    Unmod,
    ModifiedAutomodProperties,
    Ban,
    Unban,
    AutomodRejected,
    AddPermittedTerm,
    DeletePermittedTerm,
    AddBlockedTerm,
    DeleteBlockedTerm,
    ApproveAutomodMessage,
    DeniedAutomodMessage,
    Raid,
    Unraid,
    Slow,
    SlowOff,
    Followers,
    FollowersOff,
    Subscribers,
    SubscribersOff,
    EmoteOnly,
    EmoteOnlyOff,
    Clear,
    R9KBeta,
    R9KBetaOff,
    Vip,
    Unvip,
    Host,
    Unhost,
    ApproveUnbanRequest,
    DenyUnbanRequest,
    DeleteNotification,

}

This is supported on crate feature pubsub only.

A command

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Delete

Delete

Given when a message is deleted with /delete <msg-id>

Timeout

Timeout

Given when a user is timed-out with /timeout <user> <time> <reason>

Untimeout

Untimeout

Given when a user is unbanned while under a timeout /untimeout <user> or /unban <user>

Mod

Mod

Given when a user is added as a moderator. /mod <user>.

See ChatModeratorActionsReply::ModeratorAdded where this is given

Unmod

Unmod

Given when a user is removed as a moderator, /unmod <user>

ModifiedAutomodProperties

Modified automod properties

Given when automod config is changed. I.e filtering changed etc

Ban

Ban

Given when a user is banned with /timeout <user> <reason>

Unban

Unban

Given when a user is unbanned with /unban <user> or /untimeout <user>

AutomodRejected

Automod message rejected

AddPermittedTerm

👎 Deprecated since 0.5.1:
This is now returned by ChannelTermsAction, will be removed in future version of twitch_api2.

Automod permitted term added

Deprecation

This is now returned by the ChannelTermsAction action.

DeletePermittedTerm

👎 Deprecated since 0.5.1:
This is now returned by ChannelTermsAction, will be removed in future version of twitch_api2.

Automod permitted term removed

Deprecation

This is now returned by the ChannelTermsAction action.

AddBlockedTerm

👎 Deprecated since 0.5.1:
This is now returned by ChannelTermsAction, will be removed in future version of twitch_api2.

Automod blocked term added

Deprecation

This is now returned by the ChannelTermsAction action.

DeleteBlockedTerm

👎 Deprecated since 0.5.1:
This is now returned by ChannelTermsAction, will be removed in future version of twitch_api2.

Automod blocked term removed

Deprecation

This is now returned by the ChannelTermsAction action.

ApproveAutomodMessage

Automod message approved

DeniedAutomodMessage

Automod message denied

Raid

Raid

Given when editor/broadcaster does /raid <channel>

Unraid

Unraid

Given when editor/broadcaster does /unraid

Slow

Slow-mode chat enabled

SlowOff

Slow-mode chat disabled

Followers

Followers-only chat enabled

FollowersOff

Followers-only chat disabled

Subscribers

Subscriber-only chat enabled

SubscribersOff

Subscriber-only chat disabled

EmoteOnly

Emote-only chat enabled

EmoteOnlyOff

Emote-only chat disabled

Clear

Chat cleared for all viewers

R9KBeta

Unique chat enabled

R9KBetaOff

Unique chat disabled

Vip

👎 Deprecated since 0.5.1:
This is now returned by VipAdded, will be removed in future version of twitch_api2.

User added as VIP

Deprecation

This is now returned by the VipAdded action. Strangely, /unvip is still valid

Unvip

User removed as VIP

Host

Channel host started

Unhost

Channel host removed

ApproveUnbanRequest

Unban Request Approved

DenyUnbanRequest

Unban Request Denied

DeleteNotification

Users own message was deleted.