Struct twitch_api2::TwitchClient

#[non_exhaustive]
pub struct TwitchClient<'a, C> where
    C: HttpClient<'a>,  {
    pub helix: HelixClient<'a, C>,
    pub tmi: TmiClient<'a, C>,
}

This is supported on crate feature client and (crate features helix or tmi) only.

Client for Twitch APIs.

Most http clients will be able to use the 'static lifetime

pub struct MyStruct {
    twitch: TwitchClient<'static, reqwest::Client>,
    token: twitch_oauth2::AppAccessToken,
}
// etc

See client for implemented clients, you can also define your own if needed.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

helix: HelixClient<'a, C>

Helix endpoint. See helix

tmi: TmiClient<'a, C>

TMI endpoint. See tmi

Implementations

impl<C: HttpClient<'static>> TwitchClient<'static, C>

pub fn new() -> TwitchClient<'static, C> where
C: Clone + ClientDefault<'static>,

Create a new TwitchClient

impl<'a, C: HttpClient<'a>> TwitchClient<'a, C>

pub fn with_client(client: C) -> TwitchClient<'a, C> where
C: Clone,

Create a new TwitchClient with an existing HttpClient