Module twitch_api2::types

Twitch types

Structs

BadgeSetId

A Badge set ID

BadgeSetIdRef

A reference to a borrowed BadgeSetId

CategoryId

A game or category ID

CategoryIdRef

A reference to a borrowed CategoryId

ChatBadgeId

A channel chat badge ID

ChatBadgeIdRef

A reference to a borrowed ChatBadgeId

DisplayName

A users display name

DisplayNameRef

A reference to a borrowed DisplayName

EmoteId

A chat Emote ID

EmoteIdRef

A reference to a borrowed EmoteId

EmoteSetId

An Emote Set ID

EmoteSetIdRef

A reference to a borrowed EmoteSetId

EmoteUrlBuilder

Builder for emote URLs.

EventSubId

An EventSub Subscription ID

EventSubIdRef

A reference to a borrowed EventSubId

GlobalCooldown

Information about global cooldown

HypeTrainId

A Hype Train ID

HypeTrainIdRef

A reference to a borrowed HypeTrainId

Image

Links to the same image of different sizes

MsgId

A message ID

MsgIdRef

A reference to a borrowed MsgId

Nickname

A nickname, not capitalized.

NicknameRef

A reference to a borrowed Nickname

PollChoice

Poll choice

PollChoiceId

A poll choice ID

PollChoiceIdRef

A reference to a borrowed PollChoiceId

PollId

A poll ID

PollIdRef

A reference to a borrowed PollId

PredictionId

A prediction ID

PredictionIdRef

A reference to a borrowed PredictionId

PredictionOutcome

Outcome for the Prediction

PredictionOutcomeId

A prediction choice ID

PredictionOutcomeIdRef

A reference to a borrowed PredictionOutcomeId

PredictionTopPredictors

Users who were the top predictors.

RedemptionId

A reward redemption ID.

RedemptionIdRef

A reference to a borrowed RedemptionId

ResubscriptionEmote

An emote index as defined by eventsub, similar to IRC emotes twitch tag.

RewardId

A reward ID.

RewardIdRef

A reference to a borrowed RewardId

StreamId

A Stream ID

StreamIdRef

A reference to a borrowed StreamId

StreamSegmentId

A Stream Segment ID.

StreamSegmentIdRef

A reference to a borrowed StreamSegmentId

TagId

A tag ID

TagIdRef

A reference to a borrowed TagId

TeamId

A Team ID

TeamIdRef

A reference to a borrowed TeamId

Timestamp

RFC3339 timestamp

TimestampRef

A reference to a borrowed Timestamp

TwitchCategory

A game or category as defined by Twitch

User

A user according to many endpoints

UserId

A user ID.

UserIdRef

A reference to a borrowed UserId

VideoId

A video ID

VideoIdRef

A reference to a borrowed VideoId

Enums

AutomodStatus

Status of a message that is or was in AutoMod queue

BroadcasterType

Broadcaster types: “partner”, “affiliate”, or “”.

CommercialLength

Length of the commercial in seconds

CommercialLengthParseError

Error for the TryFrom on CommercialLength

EmoteAnimationSetting

Formats for an emote.

EmoteScale

Scales available for an emote.

EmoteThemeMode

Background themes available for an emote.

Max

Reward redemption max

PollStatus

Status of a poll

PredictionStatus

Status of the Prediction

SubscriptionTier

Subscription tiers

TimestampParseError

Errors that can occur when parsing a timestamp.

UserType

User types: “staff”, “admin”, “global_mod”, or “”.

VideoPeriod

Period during which the video was created

VideoPrivacy

Type of video

VideoType

Type of video

Type Definitions

UserName

A username, also specified as login. Should not be capitalized.

UserNameRef

A reference to a borrowed UserName, also specified as login. Should not be capitalized.