Struct twitch_api2::types::TimestampRef

#[repr(transparent)]
pub struct TimestampRef(_);

A reference to a borrowed Timestamp

Implementations

impl TimestampRef

pub fn from_str(raw: &str) -> Result<&Self, <Timestamp as Validator>::Error>

Transparently reinterprets the string slice as a strongly-typed TimestampRef if it conforms to Timestamp

pub unsafe fn from_str_unchecked(raw: &str) -> &Self

Transparently reinterprets the string slice as a strongly-typed TimestampRef without validating

pub fn into_owned(self: Box<TimestampRef>) -> Timestamp

Converts a Box<TimestampRef> into a Timestamp without copying or allocating

pub const fn as_str(&self) -> &str

Provides access to the underlying value as a string slice.

impl TimestampRef

pub fn normalize(&self) -> Result<Cow<'_, TimestampRef>, TimestampParseError>

Normalize the timestamp into UTC time.

Examples

use twitch_api2::types::Timestamp;

let time = Timestamp::new("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.normalize()?.as_ref(), &time);
let time2 = Timestamp::new("2021-07-01T13:37:00-01:00").unwrap();
assert_ne!(time2.normalize()?.as_ref(), &time2);

pub fn is_before<T>(&self, other: &T) -> bool where
Self: PartialOrd<T>,

Compare another time and return self < other.

Examples

use twitch_api2::types::Timestamp;

let time1 = Timestamp::new("2021-07-01T13:37:00Z").unwrap();
let time2 = Timestamp::new("2021-07-01T13:36:00Z").unwrap();
assert!(time2.is_before(&time1));
let time_now = chrono::Utc::now();
assert!(time1.is_before(&time_now));

pub fn to_day(&self) -> Timestamp

Make a timestamp with the time component set to 00:00:00.

Examples

use twitch_api2::types::Timestamp;

let time = Timestamp::new("2021-07-01T13:37:00Z").unwrap();
assert_eq!(time.to_day().as_str(), "2021-07-01T00:00:00Z")

impl TimestampRef

pub fn to_utc(&self) -> DateTime<Utc>

This is supported on crate feature chrono only.

Construct into a DateTime<Utc> time.

Panics

This method assumes the timestamp is a valid rfc3339 timestamp, and panics if not.

pub fn to_fixed_offset(&self) -> DateTime<FixedOffset>

This is supported on crate feature chrono only.

Construct into a DateTime<FixedOffset> time.

Panics

This method assumes the timestamp is a valid rfc3339 timestamp, and panics if not.