Trait twitch_api2::client::ClientDefault

pub trait ClientDefault<'a>: Clone + Sized {
    type Error: Error + Send + Sync + 'static;
    fn default_client_with_name(
        product: Option<HeaderValue>
    ) -> Result<Self, Self::Error>;

    fn default_client() -> Self { ... }
}

This is supported on crate feature client only.

A specific client default for setting some sane defaults for API calls and oauth2 usage

Associated Types

type Error: Error + Send + Sync + 'static

Errors that can happen when assembling the client

Required methods

fn default_client_with_name(
product: Option<HeaderValue>
) -> Result<Self, Self::Error>

Constructs Self with sane defaults for API calls and oauth2 and setting user-agent to include another product

Specifically, one should

Set User-Agent to {product} twitch_api2/{version_of_twitch_api2} (According to RFC7231) See TWITCH_API2_USER_AGENT for the product of this crate

Disallow redirects

Notes

When the product name is none, this function should never fail. This should be ensured with tests.

Provided methods

fn default_client() -> Self

Construct Self with sane defaults for API calls and oauth2.