Module twitch_api2::eventsub::channel::subscription

This is supported on crate feature eventsub only.

Subscription on a specified channel has changed

Modules

end

A subscription to the specified channel expires.

gift

A user gives one or more gifted subscriptions in a channel.

message

A user sends a resubscription chat message in a specific channel

Structs

ChannelSubscriptionEndV1

channel.subscription.end: a subscription to the specified channel expires.

ChannelSubscriptionEndV1Payload

channel.subscription.end response payload.

ChannelSubscriptionGiftV1

channel.subscription.gift: a subscription to the specified channel expires.

ChannelSubscriptionGiftV1Payload

channel.subscription.gift response payload.

ChannelSubscriptionMessageV1

channel.subscription.message: a subscription to the specified channel expires.

ChannelSubscriptionMessageV1Payload

channel.subscription.message response payload.