Enum twitch_api2::eventsub::EventType

#[non_exhaustive]pub enum EventType {
    ChannelUpdate,
    ChannelFollow,
    ChannelSubscribe,
    ChannelCheer,
    ChannelBan,
    ChannelUnban,
    ChannelPointsCustomRewardAdd,
    ChannelPointsCustomRewardUpdate,
    ChannelPointsCustomRewardRemove,
    ChannelPointsCustomRewardRedemptionAdd,
    ChannelPointsCustomRewardRedemptionUpdate,
    ChannelRaid,
    ChannelHypeTrainBegin,
    ChannelHypeTrainProgress,
    ChannelHypeTrainEnd,
    StreamOnline,
    StreamOffline,
    UserUpdate,
    UserAuthorizationRevoke,
}

This is supported on crate feature eventsub only.

Event name

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ChannelUpdate

channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelFollow

channel.follow: a specified channel receives a follow.

ChannelSubscribe

channel.subscribe: a specified channel receives a subscriber. This does not include resubscribes.

ChannelCheer

channel.cheer: a user cheers on the specified channel.

ChannelBan

channel.ban: a viewer is banned from the specified channel.

ChannelUnban

channel.unban: a viewer is unbanned from the specified channel.

ChannelPointsCustomRewardAdd

channel.channel_points_custom_reward.add: a custom channel points reward has been created for the specified channel.

ChannelPointsCustomRewardUpdate

channel.channel_points_custom_reward.update: a custom channel points reward has been updated for the specified channel.

ChannelPointsCustomRewardRemove

channel.channel_points_custom_reward.remove: a custom channel points reward has been removed from the specified channel.

ChannelPointsCustomRewardRedemptionAdd

channel.channel_points_custom_reward_redemption.add: a viewer has redeemed a custom channel points reward on the specified channel.

ChannelPointsCustomRewardRedemptionUpdate

channel.channel_points_custom_reward_redemption.update: a redemption of a channel points custom reward has been updated for the specified channel.

ChannelRaid

channel.raid: a broadcaster raids another broadcaster’s channel.

ChannelHypeTrainBegin

channel.hype_train.begin: a hype train begins on the specified channel.

ChannelHypeTrainProgress

channel.hype_train.progress: a hype train makes progress on the specified channel.

ChannelHypeTrainEnd

channel.hype_train.end: a hype train ends on the specified channel.

StreamOnline

stream.online: the specified broadcaster starts a stream.

StreamOffline

stream.online: the specified broadcaster stops a stream.

UserUpdate

user.update: user updates their account.

UserAuthorizationRevoke

user.authorization.revoke: a user has revoked authorization for your client id. Use this webhook to meet government requirements for handling user data, such as GDPR, LGPD, or CCPA.