Module twitch_api2::eventsub::user

This is supported on crate feature eventsub only.

Subscription types regarding users

Modules

authorization

Authorization from a user has been granted or revoked to a specific client ID

update

Specified user updates their account.

Structs

UserAuthorizationGrantV1

user.authorization.grant: a user’s authorization has been granted to your client id.

UserAuthorizationGrantV1Payload

user.authorization.grant response payload.

UserAuthorizationRevokeV1

user.authorization.revoke: a user has revoked authorization for your client id. Use this webhook to meet government requirements for handling user data, such as GDPR, LGPD, or CCPA.

UserAuthorizationRevokeV1Payload

user.authorization.revoke response payload.

UserUpdateV1

user.update: user updates their account.

UserUpdateV1Payload

user.update response payload.