Module twitch_api2::helix

This is supported on crate feature helix only.

Helix endpoints or the New Twitch API

Aside from using HelixClient as described on the crate documentation, you can decide to use this library without any specific client implementation.

use twitch_api2::helix::{self, Request, RequestGet, users::{GetUsersRequest, User}};

let request = GetUsersRequest::builder()
    .login(vec!["justintv123".to_string()])
    .build();

// Send it however you want
// Create a [`http::Response<Vec<u8>>`] with RequestGet::create_request, which takes an access token and a client_id
let response = send_http_request(request.create_request("accesstoken", "client_id")?)?;

// then parse the response
let uri = request.get_uri()?;
let user: helix::Response<_, Vec<User>> = request.parse_response(&uri,response)?;
println!("{:#?}", user);

Re-exports

pub use twitch_oauth2::Scope;

Modules

bits
Endpoints regarding bits

channels
Endpoints regarding channels

clips
Endpoints regarding clips

games
Endpoints regarding games

moderation
Endpoints regarding moderation

points
Endpoints regarding channel points/redeems

search
Endpoints regarding search

streams
Endpoints regarding streams

subscriptions
Endpoints regarding subscriptions

tags
Endpoints regarding tags

users
Endpoints regarding users

videos
Endpoints regarding videos

Structs

HelixClient	client
Client for Helix or the New Twitch API

HelixRequestPatchError
helix returned error {status:?}: {message:?} when calling PATCH {uri} with a body

HelixRequestPutError
helix returned error {status:?} - {error}: {message:?} when calling PUT {uri} with a body

Response
Response retrieved from endpoint. Data is the type in Request::Response

Enums

ClientRequestError
Errors for HelixClient::req_get and similar functions.

CreateRequestError
Could not create request

HelixRequestDeleteError
Could not parse DELETE response

HelixRequestGetError
Could not parse GET response

HelixRequestPostError
Could not parse POST response

InvalidUri
Errors that can happen when creating http::Uri for Request

SerializeError
Errors from the query serializer

Traits

Paginated
Request can be paginated with a cursor

Request
A request is a Twitch endpoint, see New Twitch API reference

RequestDelete
Helix endpoint DELETEs information

RequestGet
Helix endpoint GETs information

RequestPatch
Helix endpoint PATCHs information

RequestPost
Helix endpoint POSTs information

Type Definitions

Cursor
A cursor is a pointer to the current "page" in the twitch api pagination