Module twitch_api2::helix::bits::get_cheermotes

This is supported on crate feature helix only.

Retrieves the list of available Cheermotes, animated emotes to which viewers can assign Bits, to cheer in chat. get-cheermotes

Accessing the endpoint

Request: GetCheermotesRequest

To use this endpoint, construct a GetCheermotesRequest with the GetCheermotesRequest::builder() method. If you do not provide an ID, the request will only include global cheermotes as defined by twitch.

use twitch_api2::helix::bits::get_cheermotes;
let request = get_cheermotes::GetCheermotesRequest::builder()
    .broadcaster_id(Some("1234".into()))
    .build();
// Without broadcaster ID
let request = get_cheermotes::GetCheermotesRequest::builder().build();

Response: Cheermote

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, bits::get_cheermotes};
let request = get_cheermotes::GetCheermotesRequest::builder().build();
let response: Vec<get_cheermotes::Cheermote> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetCheermotesRequest::parse_response(None, &request.get_uri(), response)

Structs

Cheermote

Return Values for Get Cheermotes

CheermoteImage

URLs to the same image, animated or static.

CheermoteImageArray

Images in sizes

CheermoteImages

Images for different themes

GetCheermotesRequest

Query Parameters for Get Cheermotes

Level

Level of cheermote image

Tiers

Information about tier of emote

Enums

CheermoteType

Types of cheermotes