Module twitch_api2::helix::moderation

Available on crate feature helix only.

Helix endpoints regarding moderation

Modules

add_blocked_term

Adds a word or phrase to the broadcaster’s list of blocked terms. These are the terms that broadcasters don’t want used in their chat room. add-blocked-term

ban_user

Bans a user from participating in a broadcaster’s chat room, or puts them in a timeout. ban-user

check_automod_status

Determines whether a string message meets the channel’s AutoMod requirements. check-automod-status

get_banned_events

Returns all banned and timed-out users in a channel. get-banned-events

get_banned_users

Returns all banned and timed-out users in a channel. get-banned-users

get_blocked_terms

Gets the broadcaster’s list of non-private, blocked words or phrases. These are the terms that the broadcaster or moderator added manually, or that were denied by AutoMod. get-blocked-terms

get_moderator_events

Returns a list of moderators or users added and removed as moderators from a channel. get-moderator-events

get_moderators

Returns all moderators in a channel. get-moderators

manage_held_automod_messages

Allow or deny a message that was held for review by AutoMod. manage-held-automod-messages

remove_blocked_term

Removes the word or phrase that the broadcaster is blocking users from using in their chat room. remove-blocked-term

unban_user

Removes the ban or timeout that was placed on the specified user. unban-user

Structs

AddBlockedTermBody

Body Parameters for Add Blocked Term

AddBlockedTermRequest

Query Parameters for Add Blocked Term

BanUser

Return Values for Ban User

BanUserBody

Body Parameters for Ban User

BanUserRequest

Query Parameters for Ban User

BannedEvent

Return Values for Get Banned Events

BannedUser

Return Values for Get Banned Users

BlockedTerm

A blocked term in automod

CheckAutoModStatus

Return Values for Check AutoMod Status

CheckAutoModStatusBody

Body Parameters for Check AutoMod Status

CheckAutoModStatusRequest

Query Parameters for Check AutoMod Status

GetBannedEventsRequest

Query Parameters for Get Banned Events

GetBannedUsersRequest

Query Parameters for Get Banned Users

GetModeratorEventsRequest

Query Parameters for Get Moderators Events

GetModeratorsRequest

Query Parameters for Get Moderators

ManageHeldAutoModMessagesBody

Body Parameters for Manage Held AutoMod Messages

ManageHeldAutoModMessagesRequest

Query Parameters for Manage Held AutoMod Messages

Moderator

Return Values for Get Moderators

ModeratorEvent

Return Values for Get Moderators Events

RemoveBlockedTermRequest

Query Parameters for Remove Blocked Terms

UnbanUserRequest

Query Parameters for Unban User

Enums

AutoModAction

Action to take for a message.

ManageHeldAutoModMessages

Return Values for Manage Held AutoMod Messages

RemoveBlockedTerm

Return Values for Remove Blocked Terms

UnbanUserResponse

Return Values for Unban User