Module twitch_api2::helix::points

This is supported on crate feature helix only.

Helix endpoints regarding channel points/redeems

Examples

let client = HelixClient::new();
let request = GetCustomRewardRedemptionRequest::builder()
    .broadcaster_id("274637212".to_string())
    .reward_id("92af127c-7326-4483-a52b-b0da0be61c01".to_string())
    .status(CustomRewardRedemptionStatus::Canceled)
    .build();

println!("{:?}", &client.req_get(request, &token).await?.data.get(0));

Modules

create_custom_rewards
Adds a specified user to the followers of a specified channel. create-custom-rewards

delete_custom_reward
Deletes a Custom Reward on a channel. delete-custom-reward

get_custom_reward
Returns a list of Custom Reward objects for the Custom Rewards on a channel.

get_custom_reward_redemption
Returns Custom Reward Redemption objects for a Custom Reward on a channel that was created by the same client_id.

update_custom_reward
Updates a Custom Reward created on a channel.

update_redemption_status
Updates the status of Custom Reward Redemption objects on a channel that are in the UNFULFILLED status.

Structs

CreateCustomRewardBody
Body Parameters for Create Custom Rewards

CreateCustomRewardRequest
Query Parameters for Create Custom Rewards

CustomReward
Return Values for Get Custom Reward

CustomRewardRedemption
Return Values for Get Custom Reward Redemption

DeleteCustomRewardRequest
Query Parameters for Delete CustomReward

GetCustomRewardRedemptionRequest
Query Parameters for Get Custom Reward Redemption

GetCustomRewardRequest
Query Parameters for Get Custom Reward

UpdateCustomRewardBody
Body Parameters for Update Custom Rewards

UpdateCustomRewardRequest
Query Parameters for Update Custom Rewards

UpdateRedemptionStatusBody
Body Parameters for Update Redemption Status

UpdateRedemptionStatusRequest
Query Parameters for Update Redemption Status

Enums

CustomRewardRedemptionStatus
Custom reward redemption statuses: UNFULFILLED, FULFILLED or CANCELED

DeleteCustomReward
Return Values for Delete CustomReward

UpdateRedemptionStatusInformation
FIXME: Returns an object. Return Values for Update Redemption Status

Type Definitions

CreateCustomRewardResponse
Return Values for Create Custom Rewards