Module twitch_api2::helix::points::update_redemption_status

This is supported on crate feature helix only.

Updates the status of Custom Reward Redemption objects on a channel that are in the UNFULFILLED status.

Only redemptions for a reward created by the same client_id as attached to the access token can be updated. update-redemption-status

Accessing the endpoint

Request: UpdateRedemptionStatusRequest

To use this endpoint, construct a UpdateRedemptionStatusRequest with the UpdateRedemptionStatusRequest::builder() method.

use twitch_api2::helix::points::UpdateRedemptionStatusRequest;
let request = UpdateRedemptionStatusRequest::builder()
    .broadcaster_id("274637212".to_string())
    .reward_id("92af127c-7326-4483-a52b-b0da0be61c01".to_string())
    .id("17fa2df1-ad76-4804-bfa5-a40ef63efe63".to_string())
    .build();

Body: UpdateRedemptionStatusBody

We also need to provide a body to the request containing what we want to change.

use twitch_api2::helix::points::{CustomRewardRedemptionStatus, UpdateRedemptionStatusBody};
let body = UpdateRedemptionStatusBody::builder()
    .status(CustomRewardRedemptionStatus::Canceled)
    .build();

Response: UpdateRedemptionStatusInformation

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix;
use twitch_api2::helix::points::{
    CustomRewardRedemptionStatus, UpdateRedemptionStatusBody,
    UpdateRedemptionStatusInformation, UpdateRedemptionStatusRequest,
};
let client: helix::HelixClient<'static, client::DummyHttpClient> =
    helix::HelixClient::default();
let request = UpdateRedemptionStatusRequest::builder()
    .broadcaster_id("274637212".to_string())
    .reward_id("92af127c-7326-4483-a52b-b0da0be61c01".to_string())
    .id("17fa2df1-ad76-4804-bfa5-a40ef63efe63".to_string())
    .build();
let body = UpdateRedemptionStatusBody::builder()
    .status(CustomRewardRedemptionStatus::Canceled)
    .build();
let response: UpdateRedemptionStatusInformation =
    client.req_patch(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(body, &token, &client_id) and parse the http::Response with UpdateRedemptionStatusRequest::parse_response(None, &request.get_uri(), response)

Re-exports

pub use super::CustomRewardRedemption;

Structs

UpdateRedemptionStatusBody

Body Parameters for Update Redemption Status

UpdateRedemptionStatusRequest

Query Parameters for Update Redemption Status

Enums

UpdateRedemptionStatusInformation

FIXME: Returns an object. Return Values for Update Redemption Status