Module twitch_api2::helix::subscriptions::get_broadcaster_subscriptions_events

This is supported on crate feature helix only.

Returns a list of subscription events. get-broadcaster-subscriptions-events

Notes

This endpoint seems to have been removed from the docs by mistake.

Accessing the endpoint

Request: GetBroadcasterSubscriptionsEventsRequest

To use this endpoint, construct a GetBroadcasterSubscriptionsEventsRequest with the GetBroadcasterSubscriptionsEventsRequest::builder() method.

use twitch_api2::helix::subscriptions::get_broadcaster_subscriptions_events;
let request = get_broadcaster_subscriptions_events::GetBroadcasterSubscriptionsEventsRequest::builder()
    .broadcaster_id("1234")
    .build();

Response: BroadcasterSubscriptionEvent

Send the request to receive the response with HelixClient::req_get().

use twitch_api2::helix::{self, subscriptions::get_broadcaster_subscriptions_events};
let request = get_broadcaster_subscriptions_events::GetBroadcasterSubscriptionsEventsRequest::builder()
    .broadcaster_id("1234")
    .build();
let response: Vec<get_broadcaster_subscriptions_events::BroadcasterSubscriptionEvent> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetBroadcasterSubscriptionsEventsRequest::parse_response(None, &request.get_uri(), response)

Structs

BroadcasterSubscriptionEvent

Return Values for Get Broadcaster Subscriptions Events

BroadcasterSubscriptionEventData

Event data for broadcaster events.

GetBroadcasterSubscriptionsEventsRequest

Query Parameters for Get Broadcaster Subscriptions Events

Enums

BroadcasterSubscriptionEventType

Type of event