Trait twitch_api2::helix::Request

pub trait Request: Serialize {
    type Response: DeserializeOwned + PartialEq;

    pub const PATH: &'static str;
    pub const SCOPE: &'static [Scope];
    pub const OPT_SCOPE: &'static [Scope];
    pub fn query(&self) -> Result<String, Error> { ... }

    pub fn get_uri(&self) -> Result<Uri, InvalidUri> { ... }
}

This is supported on crate feature helix only.

A request is a Twitch endpoint, see New Twitch API reference

Associated Types

type Response: DeserializeOwned + PartialEq

Response type. twitch's response will deserialize to this.

Loading content...
Associated Constants

pub const PATH: &'static str

The path to the endpoint relative to the helix root. eg. channels for Get Channel Information

pub const SCOPE: &'static [Scope]

Scopes needed by this endpoint

pub const OPT_SCOPE: &'static [Scope]

Optional scopes needed by this endpoint

Loading content...
Provided methods

pub fn query(&self) -> Result<String, Error>

Defines layout of the url parameters.

pub fn get_uri(&self) -> Result<Uri, InvalidUri>

Returns full URI for the request, including query parameters.

Loading content...