Trait twitch_api2::helix::Request

pub trait Request: Serialize {
    type Response: DeserializeOwned + PartialEq;

    const PATH: &'static str;
    const SCOPE: &'static [Scope];
    const OPT_SCOPE: &'static [Scope];
    fn query(&self) -> Result<String, Error> { ... }

    fn get_uri(&self) -> Result<Uri, InvalidUri> { ... }

    fn get_bare_uri() -> Result<Uri, InvalidUri> { ... }
}

This is supported on crate feature helix only.

A request is a Twitch endpoint, see New Twitch API reference

Associated Types

type Response: DeserializeOwned + PartialEq

Response type. twitch’s response will deserialize to this.

Associated Constants

const PATH: &'static str

The path to the endpoint relative to the helix root. eg. channels for Get Channel Information

const SCOPE: &'static [Scope]

Scopes needed by this endpoint

const OPT_SCOPE: &'static [Scope]

Optional scopes needed by this endpoint

Provided methods

fn query(&self) -> Result<String, Error>

Defines layout of the url parameters.

fn get_uri(&self) -> Result<Uri, InvalidUri>

Returns full URI for the request, including query parameters.

fn get_bare_uri() -> Result<Uri, InvalidUri>

Returns bare URI for the request, NOT including query parameters.