Module twitch_api2::helix::users

Endpoints regarding users

Examples

let client = HelixClient::new();
let req = GetUsersRequest::builder()
    .login(vec!["justinfan1337".to_string()])
    .build();

println!("{:?}", &client.req_get(req, &token).await?.data);

Modules

create_user_follows
Adds a specified user to the followers of a specified channel. create-user-follows

delete_user_follows
Deletes a specified user from the followers of a specified channel. delete-user-follows

get_users
Gets information about one or more specified Twitch users. get-users

get_users_follows
Gets information on follow relationships between two Twitch users. get-users-follows

Structs

CreateUserFollowsBody
Body Parameters for Create User Follows

CreateUserFollowsRequest
Query Parameters for Create User Follows

DeleteUserFollowsRequest
Query Parameters for Delete Users Follows

GetUsersFollowsRequest
Query Parameters for Get Users Follows

GetUsersRequest
Query Parameters for Get Users

User
Return Values for Get Users

UsersFollows
Return Values for Get Users Follows

Enums

CreateUserFollows
Return Values for Create User Follows

DeleteUserFollows
Return Values for [Delete Users Follows