Module twitch_api2::pubsub

This is supported on crate feature pubsub only.

Holds serializable pubsub stuff

Use TopicSubscribe::to_command to send subscription listen and parse the responses with Response::parse

Notes

If you find that a pubsub topic reply has a field that has not yet been added to this crate, and you don't need that field, you can enable the allow_unknown_fields feature for this crate in your Cargo manifest to ignore it (and other) fields.

Undocumented features

This crate has some pubsub topics that are not documented by twitch. These may stop working at any time. To enable these, use feature unsupported to use them. Note that this crate doesn't try to keep changes to these pubsub topics semver compatible.

Modules

channel_bits
PubSub messages for bits

channel_bits_badge
PubSub messages for bits

channel_cheer	unsupported
PubSub messages for cheer events

channel_points
PubSub messages for channel points

channel_sub_gifts	unsupported
PubSub messages for sub gifts

channel_subscriptions
PubSub messages for subscriptions

community_points	unsupported
PubSub messages for community points.

following	unsupported
PubSub messages for follows

hypetrain	unsupported
PubSub messages for hype-trains

moderation
PubSub messages for moderator actions

raid	unsupported
PubSub messages for raids

video_playback	unsupported
PubSub messages for (live) stream playback information

Structs

TwitchResponse
Response from twitch PubSub

Enums

Response
Response from twitchs PubSub server. Either a response indicating status of something or a message from a topic

TopicData
Message response from twitch PubSub.

TopicSubscribe
Command that can be serialized to be sent to twitchs PubSub server to subscribe or unsubscribe to a Topic

Traits

Topic
A logical partition of messages that clients may subscribe to, to get messages.