Module twitch_api2::pubsub::hypetrain

This is supported on crate features pubsub and unsupported only.

PubSub messages for hype-trains

Structs

BitsRewards
Rewards

ConductorRewards
Reward given to conductors

Conductors
Unknown

Config
Configuration of hype train

HypeTrainApproaching
Hype train is approaching

HypeTrainConductorUpdate
Hype train conductor updated

HypeTrainEnd
Hype train ended

HypeTrainEventsV1
A user redeems an reward using channel points.

HypeTrainEventsV1Rewards
A user redeems an reward using channel points.

HypeTrainLevelUp
Hype train leveled up

HypeTrainProgress
Progress of hype train

HypeTrainProgression
Hype train progression. Akin to Participations

HypeTrainRewards
Hype train rewards

HypeTrainStart
Hype train started in channel

Kickoff
How many support events needed to start Hype Train

Level
Description of a hype-train level

NotificationThresholds
Thresholds for notifications

ParticipationConversionRates
Conversion table of event to participation points

Participations
Participations in hype train

SubsRewards
Rewards

Enums

ActionType
The specific action that was used

EndingReason
Reason hype-train ended

HypeTrainDifficulty
Difficulty of Hype Train

HypeTrainEventsV1Reply
Reply from HypeTrainEventsV1 or HypeTrainEventsV1Rewards

Reward
A reward

SourceType
Source type, same as helix hype-train events type